requirements_railway.txt
# Core dependencies for Railway deployment
python-telegram-bot>=20.7
openai>=1.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Optional dependencies
anthropic>=0.8.0
psutil>=5.9.0

# Web server for health checks
flask>=3.0.0